Small test fix

+++ torchtext/utils.py
@@ -43,6 +43,9 @@ def download_from_url(url, path):
     def process_response(r):
         chunk_size = 16 * 1024
         total_size = int(r.headers.get('Content-length', 0))
+        download_dir = os.path.dirname(path)
+        if not os.path.exists(download_dir):
+            raise RuntimeError("Download directorty {} does not exist. Did you create it?".format(download_dir))
         with open(path, "wb") as file:
             with tqdm(total=total_size, unit='B',
                       unit_scale=1, desc=path.split('/')[-1]) as t:
